move pages into pages dir + add page layout w navbar
