+++ src/App.tsx
@@ -1,13 +1,18 @@
 import { Route, Routes } from "react-router-dom";
-import SignupPage from "./components/SignUpPage/SignUpPage";
-import LoginPage from "./components/LoginPage/loginpage";
+import SignupPage from "./pages/SignUpPage/SignUpPage";
+import LoginPage from "./pages/LoginPage/LoginPage";
+import HomePage from "./pages/HomePage/HomePage";
 import "./App.css";
+import PageLayout from "./components/PageLayout";
 
 function App() {
   return (
     <Routes>
-      <Route path="/login" element={<LoginPage />} />
-      <Route path="/signup" element={<SignupPage />} />
+      <Route path="/" element={<PageLayout />}>
+        <Route index element={<HomePage />}/>
+        <Route path="login" element={<LoginPage />} />
+        <Route path="signup" element={<SignupPage />} />
+      </Route>
     </Routes>
   );
 }
